# api_documentation.yaml
openapi: 3.0.0
info:
  title: Dynamic User Flow API
  description: API for managing users with dynamic form data and sending invitations via email.
  version: 1.0.0
paths:
  /add_user:
    post:
      summary: Create a new user
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                project_id:
                  type: string
                company_name:
                  type: string
                first_name:
                  type: string
                last_name:
                  type: string
                mobile_number:
                  type: string
                email:
                  type: string
                password:
                  type: string
                hashtag:
                  type: string
                dob:
                  type: string
                  format: date
      responses:
        '200':
          description: User added successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: integer
                  user_data:
                    type: object
  /get_user/{user_id}:
    get:
      summary: Get user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '404':
          description: User not found
  /get_users:
    get:
      summary: Get all users
      responses:
        '200':
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
  /update_user/{user_id}:
    put:
      summary: Update user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
        - name: project_id
          in: formData
          required: true
          schema:
            type: string
        - name: company_name
          in: formData
          schema:
            type: string
        - name: first_name
          in: formData
          required: true
          schema:
            type: string
        - name: last_name
          in: formData
          required: true
          schema:
            type: string
        - name: mobile_number
          in: formData
          schema:
            type: string
        - name: email
          in: formData
          schema:
            type: string
        - name: password
          in: formData
          schema:
            type: string
        - name: hashtag
          in: formData
          schema:
            type: string
        - name: dob
          in: formData
          schema:
            type: string
            format: date
      responses:
        '200':
          description: User updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: integer
                  user_data:
                    type: object
        '404':
          description: User not found
  /delete_users/{user_id}:
    delete:
      summary: Delete user by ID
      parameters:
        - name: user_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: User deleted successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user_id:
                    type: integer
  /send-invitation/:
    post:
      summary: Send invitation email
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              type: object
              properties:
                to_email:
                  type: string
      responses:
        '200':
          description: Invitation email sent successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
        project_id:
          type: string
        company_name:
          type: string
        first_name:
          type: string
        last_name:
          type: string
        mobile_number:
          type: string
        email:
          type: string
        password:
          type: string
        hashtag:
          type: string
        dob:
          type: string
          format: date